Fix settlement overflow on transferred amount
Added a commit description post-PR review, as suggested.
Description is taken from https://github.com/raiden-network/raiden-contracts/pull/123#pullrequestreview-128277844 and is accurate.

This is @hackaugusto 's analysis of this solution:

First I'm going to give name to the expressions we have in the code:

D1 is the deposit for party 1
D2 is the deposit for party 2
Dc is the deposit for the channel, D1 + D2
W1 is the withdraw of party 1
W2 is the withdraw of party 2
Da is the total available deposit in the channel, D1 + D2 - W1 - W2
T1 is the transferred amount for party 1
T2 is the transferred amount for party 2
A1n is the netted amount for party 1 A1n = T2 - T1
A1r is the amount received by the party 1 A1n + D1
A1a is the available balance for party 1, A1r - W1
A2a is the available balance for party 2, Da - A1a
A1l is the locked amount for party 1
A2l is the locked amount for party 2
A1w is the withdraw amount after settle for party 1
A2w is the withdraw amount after settle for party 2

Here are the properties enforced by the smart contract:

D1, D2, and Dc cannot overflow, because D1 + D2 == Dc and Dc <= UINT256MAX
W1 and W2 cannot overflow because W1 + W2 <= Dc and Dc <= UINT256MAX
Da cannot overflow/underflow because of the above, where Da == Dc - Wc, Wc = W1 + W2 and W1 + W2 <= Dc

Here is the analysis for the happy case, where both valid and latest balance proofs are provided:

T1 and T2 overflow checks must be enforced by the Raiden clients, and we assume they are properly validated, this implies that 0 <= T1 - T2 <= D2 and 0 <= T2 - T1 <= D1
A1n = T2 - T1 cannot underflow, because this is enforced T2 >= T1. Note that with the above constraint, this means A1n <= D2, so A1n is at most D2
A1r = A1n + D1 cannot overflow, considering the worst value for A1n which is D2, we have A1r = D2 + D1 and D1 + D2 <= UINT256MAX
The withdraw values W1 and W2 are agreed and validated by the Raiden client, where the constraint W1 <= A1r and W2 <= A2r must be enforced
A1a = A1r - W1 cannot underflow, because of the constraint W1 <= A1r above
The lock values A1l and A2l are agreed and validated by the Raiden client, where the constraint A1l <= A1a and A2l <= A1a must be enforced.

Here is the analysis for the UNhappy case, where either balance proofs is not provided or the provided BP is out-dated. This still assumes the honest client properly follows the off-chain protocol:

T1 and T2 overflow checks must be enforced by the Raiden clients, and we assume they are properly validated, this implies that 0 <= T1 - T2 <= D2 and 0 <= T2 - T1 <= D1
A1n = T2 - T1 cannot underflow, because this is enforced T2 >= T1. Note that even the honest client enforcing the above constraint, the provided balance proofs could be wrong, therefor we can not say that A1n <= D2, meaning A1n can be larger than D2
A1r /can overflow/, because A1r = A1n + D1, and A1n may be larger than D2. This is okay because the party that lied allowed the honest party to receive more tokens than it has right to, effectively penalizing the liar. This overflow is handled by setting A1r to UINT256MAX, which is okay, because Dc <= UINT256MAX, meaning that a value larger than this would be meaningless.
The withdraw values W1 and W2 are agreed and validated by the Raiden client, where the constraint W1 <= A1r and W2 <= A2r must be enforced
A1a /can underflow/, if neither party provided a balance proof A1r = D1, and the only constraint used is W1 + W2 <= D1 + D2, meaning that W1 can be larger than D1. This is okay, because it will only happen if both parties lied, and we have no guarantees for badly behaving nodes. If one of the parties is honest, then the provided BP was valid, and the above constraint W1 <= A1r was enforced off-chain, and the computation wont underflow.
The lock values A1l and A2l are agreed and validated by the Raiden client, where the constraint A1l <= A1a and A2l <= A1a must be enforced.