Fix locksroot & locked amount storage after balance proofs are updated
Previously, the `locksroot` was not linked to the participant address that it belonged to. This allowed any of the two participants to claim the locked token amounts corresponding to that `locksrooot`.
This commit is simplifying the storage of the data needed to unlock the pending transfers and links the `locksroot` to a participant's address, by just using:

```
mapping(bytes32 => uint256) locksroot_identifier_to_locked_amount;
```
where the key is `keccak256(abi.encodePacked(channel_identifier, participant, locksroot));`
and the value is the `locked_amount`

The business logic is: `locksroot` is the merkle tree root of the pending transfers that `participant` has sent to his channel partner, where the channel is referenced by `channel_identifier`
This way, it does not matter who sends the `unlock` transaction, because the smart contract is aware of who should receive:
- the unlocked tokens (locks with secrets revealed on-chain) -> the partner will receive them
- the tokens corresponding to locks that don't have secrets revealed on-chain -> they will be returned to the participant