Fix contract compilation during initial install
The order in which packages are installed is not guranteed by pip / setuptools
therefore we need to include `py-solc` in `setup_requires` to make sure
it's available during the initial installation / wheel building.
Otherwise it can lead to situations where the `contracts.json.gz`
can't be built even though all dependencies are available.