Do not store '/home/yh/...' in contracts.json
Previously, contracts.json contained some absolute paths with user names
and so on. This caused verification failures after a user compiles a
contract locally. The absolute paths were used in the contract metadata
and were included in the metadata hash at the end of the deployed code.

This was one cause of
https://github.com/raiden-network/raiden-contracts/issues/408