Make sure that non-existent version is not usable
Before this commit, when the user specified the latest version
explicitly, raiden-contracts automatically chose the latest 'data'.

This commit removes that behavior.  If the user wants to use the latest
version (without caring what version it is), the user should say
`precompiled_data_path(None)` with `None` as the contract version.

This closes #606.