Remove contract_version from each contract
In order to prevent deploying multiple SecretRegistry contract on the
same chain, we need to use the same contract deployment in multiple
`contracts_version`s.

So far, every contract had a public state variable contract_version
that changes in sync.  However, since we are going to use the same
SecretRegistry with multiple versions of other contracts, we can no
longer expect `contract_version`s to move in sync.

We can try to accomodate the situation by adding special code to deal with
plural `contract_version`s but the simpler approach is to remove the
`contract_version`s from the Solidity files.

If some new identifier is needed, that will be introduced.