TokenNetwork.unlock() works with no locks
The new execution path allows a receiver to give back the
non-unlockable transfers back to the sender, when the receiver
has accidentially submitted an empty locksroot.

That's not supposed to happen with perfectly working clients,
but it's nice to remove one possibility where tokens are held
locked forever.

The new execution path doesn't allow any participants to
get more tokens than they deserve.

This fixes #1063.