Separate FAKE_ADDRESS and NOT_ADDRESS
Before this commit, a string FAKE_ADDRESS was sometimes used
as a random address and sometimes as an oddly-sized string
that's not an address.  The real value of FAKE_ADDRESS was
an oddly-sized string that's not an address.

This commit separates this constant into two: FAKE_ADDRESS and
NOT_ADDRESS. FAKE_ADDRESS is a random address while NOT_ADDRESS
is an oddly-sized string that's not an address.

This closes #1166.