Revert data to data_0.25.2
We have deployed our contracts to the mainnet and we are no longer
making frequent deployments.  But we have to be ready to
patch the Solidity contracts.  So our current Solidity sources
must be what have been deployed.

data_0.25.2 was once inconsistent #1298 partially because
we had edited the Solidity files in `data` directory.

This closes #1300.