Use solium to check all solidity files
To make it pass, some excludes have been added. Solium is only used if
it is already installed locally.

I removed the fallback from `HumanStandardToken` because I did not know
how to make it pass solium and I don't see why we would want that
behavior and it's also not in HumanStandardToken at
https://github.com/ConsenSys/Token-Factory/blob/master/contracts/HumanStandardToken.sol.

See https://github.com/raiden-network/raiden-contracts/issues/1144