Allow settling multiple times
The paid out amounts are saved in `withdrawn_amount`, now. This could
cause race conditions, if we also allow normal withdraws. This could be
solved by using two separate monotonic variables for withdraws and
settlements.

I'm hesitant to add another field right now to not increase the contract
size any further (see
https://github.com/raiden-network/raiddit/issues/39). I also expect us
to not use withdraws in the PoC, so we can do that later.