Increase the participants withdrawn_amount on unlock
To avoid overpaying participants for replayed unlock calls in a scenario

    Claim_A1(y), Claim_B1(z)
    A->B: A_LockedTransfer(x) --> BP_A-B_1(locked_amount=x, lock=X)
    registerSecret(X, X')
    B: close(Claim_B, BP_A-B_1)
    B: settle() -> payout A:y-x, B:z
    B: unlock(X) --> payout B:x
    Claim_A2(s), Claim_B2(t)
    B: close(Claim_B2, BP_A-B_1)
    B: settle() -> payout A:s-(y-x), B:t-z
    B: unlock(X) --> payout B:x

we need to make sure, that the `withdrawn_amount` values for the channel
participants are updated. This will ensure, that the second settle
payout to `B` will be reduced by `x`.

Ideally, `B` will also receive an unlocked balance proof from `A` at
some point in time, with the locked amount reduced by `x`, and the total
`transferred_amount` increased by `x` (which will then be also in sync
with the on-chain `withdrawn_amount`).