use timestamps in monitoring instead of block numbers
The utilities to generate fake test data for unlocked pending transfers was
still using block numbers. As these got replaced with timestamps, a bunch of
unit tests were still failing.

do not use named arguments

fix iou test

do not use named arguments for channel withdraw tests

adopt tests to use time travel instead of blocks

Due to the change in the contracts to measure delays with actual time instead of
blocks, it is necessary to adapt all tests to use time travel instead of mining
extra blocks. Though most test cases where still passing, this is logically
incorrect and rather luck.
There is an issue that sometimes it is necessary to add 2 more seconds instead
of just one to pass a delays window. So far this just gets accepted.

fix gas calculation test cases

Due to the recent changes that require tests to use time travel instead of
mining empty blocks, the tests used to calculate the gas usage did not work
anymore.
The most complex change was the calculation of the time delay for a monitor
transaction. So far there was just this magic number of blocks to mine. As this
is (and kinda was) more complicated now, the test case mimics the computation of
the contract. It uses the same variables and instructions except a few language
specific differences. This should keep it simple and easy to maintain.

Note that the tests do actually not 100% work yet as there are still changes on
the user deposit contract in work. Though the gas tests have been already
adapted so they should just work once this change is in place.

Having the tests fixed the `gas.json` file could be already updated. This was
done with a little "hack" to circumvent the above mentioned issue. It should be
double checked in the end if the numbers are actually correct.

fix user deposit contracts

remove last mine block references

update gas file