adapt deployment tools to work with Arbitrum
Chains that are bases on Arbitrum require special handling for gas as the
mechanisms work slightly different.

This also includes a fix for the correct order of decorator mocks in the tests.
The passed arguments of the test function are in reverse order of the decorators
creating them. In the past this has not been detected as there were to few of
them, so the order did not matter or the asserts where always equal for both
objects. This was no more true and had to be fixed.