improve gas handling for Arbitrum based networks
First of all, it is not necessary to programmatically set a fixed gas price
value when detecting an Arbitrum network. The gas price gets already set by
a commandline parameter and does not need any further doing. The default 5 Gwei
are enough for the current prices (test network). Else the user can/must set the
value higher.
As the deployment tools are meant to run by hand and usually a developer of the
team, it is not necessary to add any special gas price strategies as we do for
the Raiden services. But to support the correct/better usage, the deployment
documentation got extended for Arbitrum.