move user deposit constant into constructor
The `withdraw_timeout` is a constant value in the `UserDeposit` contract.
Recently its meaning has changed from block numbers to time in seconds. This
brings new challenges to test environments. While is was fairly easy to mine
a couple of empty blocks, manipulating time is a different story. Therefore is
has been decided to make this value configurable during deployment. This works
the same as it does for the `settle_timeout` on the `TokenNetwork` and
`TokenNetworkRegistry` contracts.